refractored types
